scanner-backend/requirements.txt
python-nmap>=0.7.1
requests>=2.31.0
rich>=13.7.0
tqdm>=4.66.1
urllib3>=2.0.7
flask>=3.0.0
flask-cors>=4.0.0

